Refactoring. Added model functions to switch between different schemes
