+++ js/model/form-model.js
@@ -1,0 +1,10 @@
+//Get needed type of model to wrok with given form scheme
+function FormModel(definition) {
+    var model = typeof definition.scheme !== 'undefined' ?
+        new LiveContractFormModel() :
+        new LegalFormModel();
+
+    this.getModel = function() {
+        return model;
+    };
+}
